tests: update for new uapi
