+++ tests/new_migration.c
@@ -1,5 +1,5 @@
-#define _GNU_SOURCEunpi
-#include "../src/common.h"
+#define _GNU_SOURCE
+#include "../include/uapi/linux/uintr.h"
 #include <errno.h>
 #include <fcntl.h>
 #include <pthread.h>
@@ -13,8 +13,6 @@
 #include <unistd.h>
 #include <x86intrin.h>
 
-static int uintr_fd = -1;
-
 /* volatile because value will change when interrupted */
 static volatile int interrupt_received = 0;
 static volatile sig_atomic_t keep_running = 1;
@@ -27,13 +25,6 @@ static void cleanup(void) {
 
   /* Disable interrupts before cleanup */
   _clui();
-
-  if (uintr_fd >= 0) {
-    printf("Unregistering handler...\n");
-    ioctl(uintr_fd, UINTR_UNREGISTER_HANDLER);
-    close(uintr_fd);
-    uintr_fd = -1;
-  }
 
   if (handler_stack) {
     free(handler_stack);
@@ -52,16 +43,49 @@ static void sigint_handler(int signum) {
   keep_running = 0;
 }
 
+/* Helper function to set thread affinity */
+static int set_thread_affinity(pthread_t thread, int core) {
+  cpu_set_t cpuset;
+  CPU_ZERO(&cpuset);
+  CPU_SET(core, &cpuset);
+
+  int ret = pthread_setaffinity_np(thread, sizeof(cpu_set_t), &cpuset);
+  if (ret != 0) {
+    printf("Failed to set thread affinity to core %d: %s\n", core,
+           strerror(ret));
+    return -1;
+  }
+  return 0;
+}
+
 void *sender_thread(void *arg) {
-  int uipi_index = *(int *)arg;
+  uintr_receiver_id_t receiver_id = *(uintr_receiver_id_t *)arg;
+  int idx = -1;
+
+  _stui()
+
+      // Set sender thread affinity to core 2
+      if (set_thread_affinity(pthread_self(), 2) != 0) {
+    return NULL;
+  }
 
   int cpu = sched_getcpu();
-  printf("Sender thread initialized on core: %d \n", cpu);
+  printf("Sender thread initialized on core: %d, pid: %d\n", cpu, getpid());
+
+  _uintr_sender_args sender_args = {
+      .receiver_id = receiver_id,
+      .flags = 0,
+      .vector = 0,
+  };
 
   sleep(3);
 
+  idx = uintr_register_sender(sender_args);
+
+  sleep(10);
+
   printf("Sending user interrupt...\n");
-  _senduipi(uipi_index);
+  _senduipi(idx);
   printf("User interrupt sent...\n");
 
   return NULL;
@@ -77,7 +101,10 @@ int main(void) {
   act.sa_handler = sigint_handler;
   sigaction(SIGINT, &act, NULL);
 
-  atexit(cleanup);
+  // Set main thread affinity to core 0
+  if (set_thread_affinity(pthread_self(), 0) != 0) {
+    return EXIT_FAILURE;
+  }
 
   // allocate a dedicated stack for the uintr handler
   handler_stack = aligned_alloc(4096, HANDLER_STACK_SIZE);
@@ -88,26 +115,14 @@ int main(void) {
   printf("Allocated handler stack at %p with size %d bytes\n", handler_stack,
          HANDLER_STACK_SIZE);
 
-  // Open the device
-  uintr_fd = open("/dev/uintr", O_RDWR);
-  if (uintr_fd < 0) {
-    perror("Failed to open /dev/uintr");
-    return EXIT_FAILURE;
-  }
-
   // Register interrupt handler
   struct _uintr_handler_args handler_args = {.handler = test_handler,
                                              .stack = handler_stack,
                                              .stack_size = HANDLER_STACK_SIZE,
                                              .flags = 0};
 
-  printf("Registering handler...\n");
-  int uipi_index = ioctl(uintr_fd, UINTR_REGISTER_HANDLER, &handler_args);
-  if (uipi_index < 0) {
-    perror("Failed to register handler");
-    goto cleanup;
-  }
-  printf("Got assigned index %d from registering handler.\n", uipi_index);
+  uintr_receiver_id_t receiver_id = 0;
+  receiver_id = uintr_register_handler(handler_args);
 
   // Enable user interrupts
   printf("Current UIF before stui: %u\n", _testui());
@@ -119,26 +134,23 @@ int main(void) {
   }
   printf("UIF set successfully. UIF after stui: %u\n", _testui());
 
-  ret = pthread_create(&sender_tid, NULL, sender_thread, &uipi_index);
+  ret = pthread_create(&sender_tid, NULL, sender_thread, &receiver_id);
   if (ret != 0) {
     perror("Failed to create sender thread");
     goto cleanup;
   }
 
-  printf("Logging MSR state main thread.\n");
-  ioctl(uintr_fd, UINTR_DUMP_MSR, cpu);
-  printf("Logging MSR state main thread.\n");
-  ioctl(uintr_fd, UINTR_DUMP_ENTRY, uipi_index);
-
+  sleep(2);
   cpu = sched_getcpu();
   printf("Main thread running on core: %d\n", cpu);
+  set_thread_affinity(pthread_self(), 8);
+  sleep(4);
+  cpu = sched_getcpu();
+  printf("Main thread migrated to core: %d\n", cpu);
   printf("Waiting for interrupt...\n");
 
   while (!interrupt_received && keep_running) {
   }
-
-  cpu = sched_getcpu();
-  printf("Interrupt received on: %d\n", cpu);
 
   if (!keep_running) {
     printf("Interrupted by user\n");
@@ -148,15 +160,10 @@ int main(void) {
   printf("User interrupt received!\n");
   pthread_join(sender_tid, NULL);
   printf("Test completed successfully!\n");
-
   ret = EXIT_SUCCESS;
 
 cleanup:
   _clui();
 
-  if (uintr_fd >= 0) {
-    ioctl(uintr_fd, UINTR_UNREGISTER_HANDLER, uipi_index);
-    close(uintr_fd);
-  }
   return ret;
 }
